requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
rfdetr
pyyaml
segment_anything
ultralytics